+++ src/ppom-controller.test.ts
@@ -633,7 +633,7 @@ describe('PPOMController', () => {
         deriveStateFromMetadata(
           controller.state,
           controller.metadata,
-          'anonymous',
+          'includeInDebugSnapshot',
         ),
       ).toMatchInlineSnapshot(`Object {}`);
     });